Minor fixes

+++ STM32_G474RE/Core/Inc/hdlc_protocol.h
@@ -23,9 +23,9 @@ typedef void (*frame_handler_type)(uint8_t *framebuffer, uint8_t framelength);
  * @param
  * @param max_frame_length
  */
-void uHdlcProtInit(sendchar_type, frame_handler_type, uint8_t max_frame_length);
+void uHdlcProtInit(sendchar_type, frame_handler_type, uint16_t max_frame_length);
 void vCharReceiver(uint8_t data);
-void vSendFrame(const uint8_t *framebuffer, uint8_t frame_length);
+void vSendFrame(const uint8_t *framebuffer, uint16_t frame_length);
